--- a/be/test/exprs/agg/aggregate_test.cpp
+++ b/be/test/exprs/agg/aggregate_test.cpp
@@ -1355,42 +1355,137 @@ TEST_F(AggregateTest, test_percentile_cont) {
 
     const AggregateFunction* func = get_aggregate_function("percentile_cont", TYPE_DOUBLE, TYPE_DOUBLE, false);
 
+    auto data_column1 = DoubleColumn::create();
+    data_column1->append(2.0);
+    data_column1->append(3.0);
+    data_column1->append(4.0);
+
+    auto data_column2 = DoubleColumn::create();
+    data_column2->append(5.0);
+    data_column2->append(6.0);
+
     // update input column 1
     auto state1 = ManagedAggrState::create(ctx, func);
-
-    auto data_column1 = DoubleColumn::create();
-    auto const_colunm1 = ColumnHelper::create_const_column<TYPE_DOUBLE>(0.1, 1);
-    data_column1->append(3.0);
-    data_column1->append(3.0);
-
+    auto const_colunm1 = ColumnHelper::create_const_column<TYPE_DOUBLE>(0.25, 1);
     std::vector<const Column*> raw_columns1;
     raw_columns1.resize(2);
     raw_columns1[0] = data_column1.get();
     raw_columns1[1] = const_colunm1.get();
-
     func->update_batch_single_state(local_ctx.get(), data_column1->size(), raw_columns1.data(), state1->state());
 
     // update input column 2
     auto state2 = ManagedAggrState::create(ctx, func);
-
-    auto data_column2 = DoubleColumn::create();
-    auto const_colunm2 = ColumnHelper::create_const_column<TYPE_DOUBLE>(0.1, 1);
-    data_column2->append(6.0);
-
+    auto const_colunm2 = ColumnHelper::create_const_column<TYPE_DOUBLE>(0.25, 1);
     std::vector<const Column*> raw_columns2;
     raw_columns2.resize(2);
     raw_columns2[0] = data_column2.get();
     raw_columns2[1] = const_colunm2.get();
-
     func->update_batch_single_state(local_ctx.get(), data_column2->size(), raw_columns2.data(), state2->state());
 
-    // merge column 1 and column 2
-    ColumnPtr serde_column = BinaryColumn::create();
-    auto result_column = DoubleColumn::create();
-    func->serialize_to_column(local_ctx.get(), state1->state(), serde_column.get());
-    func->merge(local_ctx.get(), serde_column.get(), state2->state(), 0);
-    func->finalize_to_column(local_ctx.get(), state2->state(), result_column.get());
+    // state3
+    auto state3 = ManagedAggrState::create(ctx, func);
 
+    // merge column 1 and column 2
+    auto result_column = DoubleColumn::create();
+    ColumnPtr serde_column1 = BinaryColumn::create();
+    func->serialize_to_column(local_ctx.get(), state1->state(), serde_column1.get());
+    ColumnPtr serde_column2 = BinaryColumn::create();
+    func->serialize_to_column(local_ctx.get(), state2->state(), serde_column2.get());
+
+    func->merge(local_ctx.get(), serde_column1.get(), state3->state(), 0);
+    func->merge(local_ctx.get(), serde_column2.get(), state3->state(), 0);
+
+    func->finalize_to_column(local_ctx.get(), state3->state(), result_column.get());
+
+    // [2,3,4,5,6], rate = 0.25 -> 3
+    ASSERT_EQ(3, result_column->get_data()[0]);
+}
+
+TEST_F(AggregateTest, test_percentile_cont_1) {
+    std::vector<FunctionContext::TypeDesc> arg_types = {
+            AnyValUtil::column_type_to_type_desc(TypeDescriptor::from_logical_type(TYPE_DOUBLE)),
+            AnyValUtil::column_type_to_type_desc(TypeDescriptor::from_logical_type(TYPE_DOUBLE))};
+    auto return_type = AnyValUtil::column_type_to_type_desc(TypeDescriptor::from_logical_type(TYPE_DOUBLE));
+    std::unique_ptr<FunctionContext> local_ctx(FunctionContext::create_test_context(std::move(arg_types), return_type));
+
+    const AggregateFunction* func = get_aggregate_function("percentile_cont", TYPE_DOUBLE, TYPE_DOUBLE, false);
+
+    auto data_column1 = DoubleColumn::create();
+    data_column1->append(2.0);
+    data_column1->append(3.0);
+    data_column1->append(4.0);
+
+    auto data_column2 = DoubleColumn::create();
+    data_column2->append(5.0);
+    data_column2->append(6.0);
+
+    // update input column 1
+    auto state1 = ManagedAggrState::create(ctx, func);
+    auto const_colunm1 = ColumnHelper::create_const_column<TYPE_DOUBLE>(1, 1);
+    std::vector<const Column*> raw_columns1;
+    raw_columns1.resize(2);
+    raw_columns1[0] = data_column1.get();
+    raw_columns1[1] = const_colunm1.get();
+    func->update_batch_single_state(local_ctx.get(), data_column1->size(), raw_columns1.data(), state1->state());
+
+    // update input column 2
+    auto state2 = ManagedAggrState::create(ctx, func);
+    auto const_colunm2 = ColumnHelper::create_const_column<TYPE_DOUBLE>(1, 1);
+    std::vector<const Column*> raw_columns2;
+    raw_columns2.resize(2);
+    raw_columns2[0] = data_column2.get();
+    raw_columns2[1] = const_colunm2.get();
+    func->update_batch_single_state(local_ctx.get(), data_column2->size(), raw_columns2.data(), state2->state());
+
+    // state3
+    auto state3 = ManagedAggrState::create(ctx, func);
+
+    // merge column 1 and column 2
+    auto result_column = DoubleColumn::create();
+    ColumnPtr serde_column1 = BinaryColumn::create();
+    func->serialize_to_column(local_ctx.get(), state1->state(), serde_column1.get());
+    ColumnPtr serde_column2 = BinaryColumn::create();
+    func->serialize_to_column(local_ctx.get(), state2->state(), serde_column2.get());
+
+    func->merge(local_ctx.get(), serde_column1.get(), state3->state(), 0);
+    func->merge(local_ctx.get(), serde_column2.get(), state3->state(), 0);
+
+    func->finalize_to_column(local_ctx.get(), state3->state(), result_column.get());
+
+    // [2,3,4,5,6], rate = 1 -> 6
+    ASSERT_EQ(6, result_column->get_data()[0]);
+}
+
+TEST_F(AggregateTest, test_percentile_cont_2) {
+    std::vector<FunctionContext::TypeDesc> arg_types = {
+            AnyValUtil::column_type_to_type_desc(TypeDescriptor::from_logical_type(TYPE_DOUBLE)),
+            AnyValUtil::column_type_to_type_desc(TypeDescriptor::from_logical_type(TYPE_DOUBLE))};
+    auto return_type = AnyValUtil::column_type_to_type_desc(TypeDescriptor::from_logical_type(TYPE_DOUBLE));
+    std::unique_ptr<FunctionContext> local_ctx(FunctionContext::create_test_context(std::move(arg_types), return_type));
+
+    const AggregateFunction* func = get_aggregate_function("percentile_cont", TYPE_DOUBLE, TYPE_DOUBLE, false);
+
+    auto data_column1 = DoubleColumn::create();
+    data_column1->append(2.0);
+    data_column1->append(3.0);
+    data_column1->append(4.0);
+    data_column1->append(5.0);
+    data_column1->append(6.0);
+
+    // update input column 1
+    auto state1 = ManagedAggrState::create(ctx, func);
+    auto const_colunm1 = ColumnHelper::create_const_column<TYPE_DOUBLE>(0.25, 1);
+    std::vector<const Column*> raw_columns1;
+    raw_columns1.resize(2);
+    raw_columns1[0] = data_column1.get();
+    raw_columns1[1] = const_colunm1.get();
+    func->update_batch_single_state(local_ctx.get(), data_column1->size(), raw_columns1.data(), state1->state());
+
+    // merge column 1 and column 2
+    auto result_column = DoubleColumn::create();
+    func->finalize_to_column(local_ctx.get(), state1->state(), result_column.get());
+
+    // [2,3,4,5,6], rate = 0.25 -> 3
     ASSERT_EQ(3, result_column->get_data()[0]);
 }