--- a/test/sql/test_agg_function/T/test_percentile_cont
+++ b/test/sql/test_agg_function/T/test_percentile_cont
@@ -1,0 +1,50 @@
+-- name: test_percentile_cont
+
+CREATE TABLE `test_pc` (
+  `date` date NULL COMMENT "",
+  `datetime` datetime NULL COMMENT "",
+  `db` double NULL COMMENT "",
+  `id` int(11) NULL COMMENT "",
+  `name` varchar(255) NULL COMMENT "",
+  `subject` varchar(255) NULL COMMENT "",
+  `score` int(11) NULL COMMENT ""
+) ENGINE=OLAP
+DUPLICATE KEY(`date`)
+DISTRIBUTED BY HASH(`id`) BUCKETS 4
+PROPERTIES (
+"replication_num" = "1",
+"in_memory" = "false",
+"enable_persistent_index" = "false",
+"replicated_storage" = "true",
+"compression" = "LZ4"
+);
+
+
+insert into test_pc values ("2018-01-01","2018-01-01 00:00:01",11.1,1,"Tom","English",90);
+insert into test_pc values ("2019-01-01","2019-01-01 00:00:01",11.2,1,"Tom","English",91);
+insert into test_pc values ("2020-01-01","2020-01-01 00:00:01",11.3,1,"Tom","English",92);
+insert into test_pc values ("2021-01-01","2021-01-01 00:00:01",11.4,1,"Tom","English",93);
+insert into test_pc values ("2022-01-01","2022-01-01 00:00:01",11.5,1,"Tom","English",94);
+insert into test_pc values (NULL,NULL,NULL,NULL,"Tom","English",NULL);
+
+
+select percentile_cont(score, 0) from test_pc;
+select percentile_cont(score, 0.25) from test_pc;
+select percentile_cont(score, 0.5) from test_pc;
+select percentile_cont(score, 0.75) from test_pc;
+select percentile_cont(score, 1) from test_pc;
+select percentile_cont(date, 0) from test_pc;
+select percentile_cont(date, 0.25) from test_pc;
+select percentile_cont(date, 0.5) from test_pc;
+select percentile_cont(date, 0.75) from test_pc;
+select percentile_cont(date, 1) from test_pc;
+select percentile_cont(datetime, 0) from test_pc;
+select percentile_cont(datetime, 0.25) from test_pc;
+select percentile_cont(datetime, 0.5) from test_pc;
+select percentile_cont(datetime, 0.75) from test_pc;
+select percentile_cont(datetime, 1) from test_pc;
+select percentile_cont(db, 0) from test_pc;
+select percentile_cont(db, 0.25) from test_pc;
+select percentile_cont(db, 0.5) from test_pc;
+select percentile_cont(db, 0.75) from test_pc;
+select percentile_cont(db, 1) from test_pc;