Why I'm doing:
percentile_cont performance enhance

What I'm doing:
Detailed information about this enhancement has been put in issue below:
https://github.com/StarRocks/starrocks/issues/36160

Test on the same cluster and the same data set, and record the performance before and after optimization.

8506500 items —— 0.54 sec --> 0.20 sec

54387164 items —— 5.15 sec --> 0.80 sec

109610515 items —— 8.65 sec --> 1.50 sec

251414166 items —— 20.10 sec --> 3.42 sec

742991422 items —— 1 min 14.80 sec --> 10.88 sec

1106646823 items —— 2 min 55.81 sec --> 16.16 sec

The test shows that the processing performance has been significantly improved.
Especially the processing performance under large data volume is close to 10 times that before optimization.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.


## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:

- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.2
  - [x] 3.1
  - [x] 3.0
  - [x] 2.5