@rubiesvelt 2023-11-30
> @rubiesvelt it would be appreciated if you can open an issue, and move the comprehensive context and background information to the issue, just keep a few brief summary here and give a reference to the issue.

Sure, detailed information about this enhancement has been put in the issue https://github.com/StarRocks/starrocks/issues/36160