@rubiesvelt 2023-11-30
> Could you paste the performance gain from this change?

No problem, performance gain has been put in the PR and detailed information about this enhancement had been put in this issue https://github.com/StarRocks/starrocks/issues/36160