@rubiesvelt 2023-12-20
> did you compare the performance with `starrocks::merge_sorted_chunks` ?

No we didn't do that comparison yet, we tried to optimize the performance to the best and did comparison with the previous approach.